move functions to helper module
