+++ localstack/services/apigateway/apigateway_listener.py
@@ -1,13 +1,10 @@
 import base64
-import datetime
 import json
 import logging
 import re
-import time
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Union
 from urllib.parse import urljoin
 
-import pytz
 import requests
 from flask import Response as FlaskResponse
 from jsonschema import ValidationError, validate
@@ -18,8 +15,6 @@ from localstack.constants import (
     APPLICATION_JSON,
     HEADER_LOCALSTACK_AUTHORIZATION,
     HEADER_LOCALSTACK_EDGE_URL,
-    LOCALHOST_HOSTNAME,
-    PATH_USER_REQUEST,
     TEST_AWS_ACCOUNT_ID,
 )
 from localstack.services.apigateway import helpers
@@ -36,7 +31,6 @@ from localstack.services.apigateway.helpers import (
     extract_path_params,
     extract_query_string_params,
     get_cors_response,
-    get_resource_for_path,
     handle_accounts,
     handle_authorizers,
     handle_base_path_mappings,
@@ -45,7 +39,7 @@ from localstack.services.apigateway.helpers import (
     handle_gateway_responses,
     handle_validators,
     handle_vpc_links,
-    make_error_response,
+    make_error_response, PATH_REGEX_USER_REQUEST,
 )
 from localstack.services.apigateway.integration import (
     RequestTemplates,
@@ -66,9 +60,7 @@ from localstack.utils.aws.aws_responses import (
     request_response_stream,
     requests_response,
 )
-from localstack.utils.aws.request_context import MARKER_APIGW_REQUEST_REGION, THREAD_LOCAL
-from localstack.utils.common import camel_to_snake_case, json_safe, long_uid, to_bytes, to_str
-
+from localstack.utils.common import camel_to_snake_case, json_safe, to_bytes, to_str
 # set up logger
 from localstack.utils.http import add_query_params_to_url
 
@@ -79,17 +71,6 @@ TARGET_REGEX_PATH_S3_URI = (
     r"^arn:aws:apigateway:[a-zA-Z0-9\-]+:s3:path/(?P<bucket>[^/]+)/(?P<object>.+)$"
 )
 TARGET_REGEX_ACTION_S3_URI = r"^arn:aws:apigateway:[a-zA-Z0-9\-]+:s3:action/(?:GetObject&Bucket\=(?P<bucket>[^&]+)&Key\=(?P<object>.+))$"
-# regex path pattern for user requests, handles stages like $default
-PATH_REGEX_USER_REQUEST = (
-    r"^/restapis/([A-Za-z0-9_\\-]+)(?:/([A-Za-z0-9\_($|%%24)\\-]+))?/%s/(.*)$" % PATH_USER_REQUEST
-)
-# URL pattern for invocations
-HOST_REGEX_EXECUTE_API = (
-    r"(?:.*://)?([a-zA-Z0-9-]+)\.execute-api\.(%s|([^\.]+)\.amazonaws\.com)(.*)"
-    % LOCALHOST_HOSTNAME
-)
-
-REQUEST_TIME_DATE_FORMAT = "%d/%b/%Y:%H:%M:%S %z"
 
 
 class AuthorizationError(Exception):
@@ -367,77 +348,8 @@ def apply_response_parameters(invocation_context: ApiInvocationContext):
     return response
 
 
-def set_api_id_stage_invocation_path(
-    invocation_context: ApiInvocationContext,
-) -> ApiInvocationContext:
-    # skip if all details are already available
-    values = (
-        invocation_context.api_id,
-        invocation_context.stage,
-        invocation_context.path_with_query_string,
-    )
-    if all(values):
-        return invocation_context
-
-    # skip if this is a websocket request
-    if invocation_context.is_websocket_request():
-        return invocation_context
-
-    path = invocation_context.path
-    headers = invocation_context.headers
-
-    path_match = re.search(PATH_REGEX_USER_REQUEST, path)
-    host_header = headers.get(HEADER_LOCALSTACK_EDGE_URL, "") or headers.get("Host") or ""
-    host_match = re.search(HOST_REGEX_EXECUTE_API, host_header)
-    test_invoke_match = re.search(PATH_REGEX_TEST_INVOKE_API, path)
-    if path_match:
-        api_id = path_match.group(1)
-        stage = path_match.group(2)
-        relative_path_w_query_params = "/%s" % path_match.group(3)
-    elif host_match:
-        api_id = extract_api_id_from_hostname_in_url(host_header)
-        stage = path.strip("/").split("/")[0]
-        relative_path_w_query_params = "/%s" % path.lstrip("/").partition("/")[2]
-    elif test_invoke_match:
-        # special case: fetch the resource details for TestInvokeApi invocations
-        stage = None
-        region_name = invocation_context.region_name
-        api_id = test_invoke_match.group(1)
-        resource_id = test_invoke_match.group(2)
-        query_string = test_invoke_match.group(4) or ""
-        apigateway = aws_stack.connect_to_service(
-            service_name="apigateway", region_name=region_name
-        )
-        resource = apigateway.get_resource(restApiId=api_id, resourceId=resource_id)
-        resource_path = resource.get("path")
-        relative_path_w_query_params = f"{resource_path}{query_string}"
-    else:
-        raise Exception(
-            f"Unable to extract API Gateway details from request: {path} {dict(headers)}"
-        )
-    if api_id:
-        # set current region in request thread local, to ensure aws_stack.get_region() works properly
-        if getattr(THREAD_LOCAL, "request_context", None) is not None:
-            THREAD_LOCAL.request_context.headers[MARKER_APIGW_REQUEST_REGION] = API_REGIONS.get(
-                api_id, ""
-            )
-
-    # set details in invocation context
-    invocation_context.api_id = api_id
-    invocation_context.stage = stage
-    invocation_context.path_with_query_string = relative_path_w_query_params
-    return invocation_context
-
-
-def extract_api_id_from_hostname_in_url(hostname: str) -> str:
-    """Extract API ID 'id123' from URLs like https://id123.execute-api.localhost.localstack.cloud:4566"""
-    match = re.match(HOST_REGEX_EXECUTE_API, hostname)
-    api_id = match.group(1)
-    return api_id
-
-
 def invoke_rest_api_from_request(invocation_context: ApiInvocationContext):
-    set_api_id_stage_invocation_path(invocation_context)
+    helpers.set_api_id_stage_invocation_path(invocation_context)
     try:
         return invoke_rest_api(invocation_context)
     except AuthorizationError as e:
@@ -454,7 +366,7 @@ def invoke_rest_api(invocation_context: ApiInvocationContext):
     # run gateway authorizers for this request
     authorize_invocation(invocation_context)
 
-    extracted_path, resource = get_target_resource_details(invocation_context)
+    extracted_path, resource = helpers.get_target_resource_details(invocation_context)
     if not resource:
         return make_error_response("Unable to find path %s" % invocation_context.path, 404)
 
@@ -553,7 +465,7 @@ def invoke_rest_api_integration_backend(invocation_context: ApiInvocationContext
                     uri.split(":lambda:path")[1].split("functions/")[1].split("/invocations")[0]
                 )
 
-            invocation_context.context = get_event_request_context(invocation_context)
+            invocation_context.context = helpers.get_event_request_context(invocation_context)
             invocation_context.stage_variables = helpers.get_stage_variables(invocation_context)
             if invocation_context.authorizer_type:
                 authorizer_context = {
@@ -638,7 +550,7 @@ def invoke_rest_api_integration_backend(invocation_context: ApiInvocationContext
                 target = ""
 
             try:
-                invocation_context.context = get_event_request_context(invocation_context)
+                invocation_context.context = helpers.get_event_request_context(invocation_context)
                 invocation_context.stage_variables = helpers.get_stage_variables(invocation_context)
                 request_templates = RequestTemplates()
                 payload = request_templates.render(invocation_context)
@@ -775,7 +687,7 @@ def invoke_rest_api_integration_backend(invocation_context: ApiInvocationContext
                 )
                 return result
             elif uri.startswith("arn:aws:apigateway:") and ":sns:path" in uri:
-                invocation_context.context = get_event_request_context(invocation_context)
+                invocation_context.context = helpers.get_event_request_context(invocation_context)
                 invocation_context.stage_variables = helpers.get_stage_variables(invocation_context)
 
                 integration_response = SnsIntegration().invoke(invocation_context)
@@ -836,7 +748,7 @@ def invoke_rest_api_integration_backend(invocation_context: ApiInvocationContext
                 uri = "http://%s/%s" % (instance["Id"], invocation_path.lstrip("/"))
 
         # apply custom request template
-        invocation_context.context = get_event_request_context(invocation_context)
+        invocation_context.context = helpers.get_event_request_context(invocation_context)
         invocation_context.stage_variables = helpers.get_stage_variables(invocation_context)
         request_templates = RequestTemplates()
         payload = request_templates.render(invocation_context)
@@ -872,83 +784,6 @@ def invoke_rest_api_integration_backend(invocation_context: ApiInvocationContext
         'API Gateway integration type "%s", method "%s", URI "%s" not yet implemented'
         % (integration_type, method, uri)
     )
-
-
-def get_target_resource_details(invocation_context: ApiInvocationContext) -> Tuple[str, Dict]:
-    """Look up and return the API GW resource (path pattern + resource dict) for the given invocation context."""
-    path_map = helpers.get_rest_api_paths(
-        rest_api_id=invocation_context.api_id, region_name=invocation_context.region_name
-    )
-    relative_path = invocation_context.invocation_path
-    try:
-        extracted_path, resource = get_resource_for_path(path=relative_path, path_map=path_map)
-        invocation_context.resource = resource
-        return extracted_path, resource
-    except Exception:
-        return None, None
-
-
-def get_target_resource_method(invocation_context: ApiInvocationContext) -> Optional[Dict]:
-    """Look up and return the API GW resource method for the given invocation context."""
-    _, resource = get_target_resource_details(invocation_context)
-    if not resource:
-        return None
-    methods = resource.get("resourceMethods") or {}
-    method_name = invocation_context.method.upper()
-    return methods.get(method_name) or methods.get("ANY")
-
-
-def get_event_request_context(invocation_context: ApiInvocationContext):
-    method = invocation_context.method
-    path = invocation_context.path
-    headers = invocation_context.headers
-    integration_uri = invocation_context.integration_uri
-    resource_path = invocation_context.resource_path
-    resource_id = invocation_context.resource_id
-
-    set_api_id_stage_invocation_path(invocation_context)
-    relative_path, query_string_params = extract_query_string_params(
-        path=invocation_context.path_with_query_string
-    )
-    api_id = invocation_context.api_id
-    stage = invocation_context.stage
-
-    source_ip = headers.get("X-Forwarded-For", ",").split(",")[-2].strip()
-    integration_uri = integration_uri or ""
-    account_id = integration_uri.split(":lambda:path")[-1].split(":function:")[0].split(":")[-1]
-    account_id = account_id or TEST_AWS_ACCOUNT_ID
-    request_context = {
-        "accountId": account_id,
-        "apiId": api_id,
-        "resourcePath": resource_path or relative_path,
-        "domainPrefix": invocation_context.domain_prefix,
-        "domainName": invocation_context.domain_name,
-        "resourceId": resource_id,
-        "requestId": long_uid(),
-        "identity": {
-            "accountId": account_id,
-            "sourceIp": source_ip,
-            "userAgent": headers.get("User-Agent"),
-        },
-        "httpMethod": method,
-        "protocol": "HTTP/1.1",
-        "requestTime": pytz.utc.localize(datetime.datetime.utcnow()).strftime(
-            REQUEST_TIME_DATE_FORMAT
-        ),
-        "requestTimeEpoch": int(time.time() * 1000),
-        "authorizer": {},
-    }
-
-    # set "authorizer" and "identity" event attributes from request context
-    auth_context = invocation_context.auth_context
-    if auth_context:
-        request_context["authorizer"] = auth_context
-    request_context["identity"].update(invocation_context.auth_identity or {})
-
-    if not helpers.is_test_invoke_method(method, path):
-        request_context["path"] = (f"/{stage}" if stage else "") + relative_path
-        request_context["stage"] = stage
-    return request_context
 
 
 def apply_request_response_templates(
