+++ localstack/services/apigateway/helpers.py
@@ -1,9 +1,12 @@
+import datetime
 import json
 import logging
 import re
+import time
 from typing import Any, Dict, List, Optional, Tuple
 from urllib import parse as urlparse
 
+import pytz
 from botocore.utils import InvalidArnException
 from jsonpatch import apply_patch
 from jsonpointer import JsonPointerException
@@ -24,8 +27,21 @@ from localstack.utils import common
 from localstack.utils.aws import aws_stack
 from localstack.utils.aws.aws_responses import requests_error_response_json, requests_response
 from localstack.utils.aws.aws_stack import parse_arn
+from localstack.utils.aws.request_context import MARKER_APIGW_REQUEST_REGION, THREAD_LOCAL
 
 LOG = logging.getLogger(__name__)
+
+REQUEST_TIME_DATE_FORMAT = "%d/%b/%Y:%H:%M:%S %z"
+
+# regex path pattern for user requests, handles stages like $default
+PATH_REGEX_USER_REQUEST = (
+    r"^/restapis/([A-Za-z0-9_\\-]+)(?:/([A-Za-z0-9\_($|%%24)\\-]+))?/%s/(.*)$" % PATH_USER_REQUEST
+)
+# URL pattern for invocations
+HOST_REGEX_EXECUTE_API = (
+    r"(?:.*://)?([a-zA-Z0-9-]+)\.execute-api\.(%s|([^\.]+)\.amazonaws\.com)(.*)"
+    % LOCALHOST_HOSTNAME
+)
 
 # regex path patterns
 PATH_REGEX_MAIN = r"^/restapis/([A-Za-z0-9_\-]+)/[a-z]+(\?.*)?"
@@ -1210,3 +1226,147 @@ def import_api_from_openapi_spec(
         rest_api.endpoint_configuration = endpoint_config
 
     return rest_api
+
+
+def get_target_resource_details(invocation_context: ApiInvocationContext) -> Tuple[str, Dict]:
+    """Look up and return the API GW resource (path pattern + resource dict) for the given invocation context."""
+    path_map = get_rest_api_paths(
+        rest_api_id=invocation_context.api_id, region_name=invocation_context.region_name
+    )
+    relative_path = invocation_context.invocation_path
+    try:
+        extracted_path, resource = get_resource_for_path(path=relative_path, path_map=path_map)
+        invocation_context.resource = resource
+        return extracted_path, resource
+    except Exception:
+        return None, None
+
+
+def get_target_resource_method(invocation_context: ApiInvocationContext) -> Optional[Dict]:
+    """Look up and return the API GW resource method for the given invocation context."""
+    _, resource = get_target_resource_details(invocation_context)
+    if not resource:
+        return None
+    methods = resource.get("resourceMethods") or {}
+    method_name = invocation_context.method.upper()
+    return methods.get(method_name) or methods.get("ANY")
+
+
+def get_event_request_context(invocation_context: ApiInvocationContext):
+    method = invocation_context.method
+    path = invocation_context.path
+    headers = invocation_context.headers
+    integration_uri = invocation_context.integration_uri
+    resource_path = invocation_context.resource_path
+    resource_id = invocation_context.resource_id
+
+    set_api_id_stage_invocation_path(invocation_context)
+    relative_path, query_string_params = extract_query_string_params(
+        path=invocation_context.path_with_query_string
+    )
+    api_id = invocation_context.api_id
+    stage = invocation_context.stage
+
+    source_ip = headers.get("X-Forwarded-For", ",").split(",")[-2].strip()
+    integration_uri = integration_uri or ""
+    account_id = integration_uri.split(":lambda:path")[-1].split(":function:")[0].split(":")[-1]
+    account_id = account_id or TEST_AWS_ACCOUNT_ID
+    request_context = {
+        "accountId": account_id,
+        "apiId": api_id,
+        "resourcePath": resource_path or relative_path,
+        "domainPrefix": invocation_context.domain_prefix,
+        "domainName": invocation_context.domain_name,
+        "resourceId": resource_id,
+        "requestId": long_uid(),
+        "identity": {
+            "accountId": account_id,
+            "sourceIp": source_ip,
+            "userAgent": headers.get("User-Agent"),
+        },
+        "httpMethod": method,
+        "protocol": "HTTP/1.1",
+        "requestTime": pytz.utc.localize(datetime.datetime.utcnow()).strftime(
+            REQUEST_TIME_DATE_FORMAT
+        ),
+        "requestTimeEpoch": int(time.time() * 1000),
+        "authorizer": {},
+    }
+
+    # set "authorizer" and "identity" event attributes from request context
+    auth_context = invocation_context.auth_context
+    if auth_context:
+        request_context["authorizer"] = auth_context
+    request_context["identity"].update(invocation_context.auth_identity or {})
+
+    if not is_test_invoke_method(method, path):
+        request_context["path"] = (f"/{stage}" if stage else "") + relative_path
+        request_context["stage"] = stage
+    return request_context
+
+def set_api_id_stage_invocation_path(
+    invocation_context: ApiInvocationContext,
+) -> ApiInvocationContext:
+    # skip if all details are already available
+    values = (
+        invocation_context.api_id,
+        invocation_context.stage,
+        invocation_context.path_with_query_string,
+    )
+    if all(values):
+        return invocation_context
+
+    # skip if this is a websocket request
+    if invocation_context.is_websocket_request():
+        return invocation_context
+
+    path = invocation_context.path
+    headers = invocation_context.headers
+
+    path_match = re.search(PATH_REGEX_USER_REQUEST, path)
+    host_header = headers.get(HEADER_LOCALSTACK_EDGE_URL, "") or headers.get("Host") or ""
+    host_match = re.search(HOST_REGEX_EXECUTE_API, host_header)
+    test_invoke_match = re.search(PATH_REGEX_TEST_INVOKE_API, path)
+    if path_match:
+        api_id = path_match.group(1)
+        stage = path_match.group(2)
+        relative_path_w_query_params = "/%s" % path_match.group(3)
+    elif host_match:
+        api_id = extract_api_id_from_hostname_in_url(host_header)
+        stage = path.strip("/").split("/")[0]
+        relative_path_w_query_params = "/%s" % path.lstrip("/").partition("/")[2]
+    elif test_invoke_match:
+        # special case: fetch the resource details for TestInvokeApi invocations
+        stage = None
+        region_name = invocation_context.region_name
+        api_id = test_invoke_match.group(1)
+        resource_id = test_invoke_match.group(2)
+        query_string = test_invoke_match.group(4) or ""
+        apigateway = aws_stack.connect_to_service(
+            service_name="apigateway", region_name=region_name
+        )
+        resource = apigateway.get_resource(restApiId=api_id, resourceId=resource_id)
+        resource_path = resource.get("path")
+        relative_path_w_query_params = f"{resource_path}{query_string}"
+    else:
+        raise Exception(
+            f"Unable to extract API Gateway details from request: {path} {dict(headers)}"
+        )
+    if api_id:
+        # set current region in request thread local, to ensure aws_stack.get_region() works properly
+        if getattr(THREAD_LOCAL, "request_context", None) is not None:
+            THREAD_LOCAL.request_context.headers[MARKER_APIGW_REQUEST_REGION] = API_REGIONS.get(
+                api_id, ""
+            )
+
+    # set details in invocation context
+    invocation_context.api_id = api_id
+    invocation_context.stage = stage
+    invocation_context.path_with_query_string = relative_path_w_query_params
+    return invocation_context
+
+def extract_api_id_from_hostname_in_url(hostname: str) -> str:
+    """Extract API ID 'id123' from URLs like https://id123.execute-api.localhost.localstack.cloud:4566"""
+    match = re.match(HOST_REGEX_EXECUTE_API, hostname)
+    api_id = match.group(1)
+    return api_id
